+++ packages/code/src/Code.tsx
@@ -1,4 +1,10 @@
-import React, { useMemo } from 'react';
+import React, {
+  useMemo,
+  useRef,
+  useEffect,
+  useLayoutEffect,
+  useState,
+} from 'react';
 import PropTypes from 'prop-types';
 import { css, cx } from '@leafygreen-ui/emotion';
 import Syntax, {
@@ -7,8 +13,13 @@ import Syntax, {
   Language,
   variantColors,
 } from '@leafygreen-ui/syntax';
+import Icon from '@leafygreen-ui/icon';
+import IconButton from '@leafygreen-ui/icon-button';
 import LineNumbers from './LineNumbers';
 import WindowChrome from './WindowChrome';
+import debounce from 'lodash/debounce';
+import { uiColors } from '@leafygreen-ui/palette';
+import ClipboardJS from 'clipboard';
 
 function stringFragmentIsBlank(str: string): str is '' | ' ' {
   return str === '' || str === ' ';
@@ -60,6 +71,7 @@ const codeWrapperStyle = css`
   padding: ${whiteSpace}px;
   margin: 0;
   position: relative;
+  flex-grow: 1;
 `;
 
 const codeWrapperStyleWithLineNumbers = css`
@@ -70,14 +82,124 @@ const codeWrapperStyleWithWindowChrome = css`
   border-left: 0;
 `;
 
+const copyStyle = css`
+  width: 38px;
+  border-left: solid 1px;
+  display: flex;
+  flex-direction: column;
+  flex-shrink: 0;
+  padding-top: 6px;
+`;
+
+const singleLineCopyStyle = css`
+  padding: 10px;
+`;
+
 function getWrapperVariantStyle(variant: Variant): string {
-  const colors = variantColors[variant];
+  const colors = variantColors[variant] ?? variantColors[Variant.Light];
 
   return css`
     border-color: ${colors[1]};
     background-color: ${colors[0]};
     color: ${colors[3]};
   `;
+}
+
+function getSidebarVariantStyle(variant: Variant): string {
+  const colors = variantColors[variant] ?? variantColors[Variant.Light];
+
+  switch (variant) {
+    case Variant.Light:
+      return css`
+        border-color: ${colors[1]};
+        background-color: white;
+      `;
+    case Variant.Dark:
+      return css`
+        border-color: ${colors[1]};
+        background-color: ${colors[1]};
+      `;
+  }
+}
+
+function getCopyButtonStyle(variant: Variant, copied: boolean): string {
+  const baseStyle = css`
+    align-self: center;
+    color: ${uiColors.gray.base};
+  `;
+
+  if (copied) {
+    return cx(
+      baseStyle,
+      css`
+        color: ${uiColors.white};
+        background-color: ${uiColors.green.base};
+        &:focus {
+          color: ${uiColors.white};
+          &:before {
+            background-color: ${uiColors.green.base};
+          }
+        }
+        &:hover {
+          color: ${uiColors.white};
+          &:before {
+            background-color: ${uiColors.green.base};
+          }
+        }
+      `,
+    );
+  }
+
+  if (variant === Variant.Dark) {
+    return cx(
+      baseStyle,
+      css`
+        background-color: ${uiColors.gray.dark3};
+      `,
+    );
+  }
+
+  return baseStyle;
+}
+
+const ScrollState = {
+  None: 'none',
+  Left: 'left',
+  Right: 'right',
+  Both: 'both',
+} as const;
+
+type ScrollState = typeof ScrollState[keyof typeof ScrollState];
+
+function getScrollShadowStyle(
+  scrollState: ScrollState,
+  variant: Variant,
+): string {
+  const colors = variantColors[variant];
+  const shadowColor =
+    variant === Variant.Light ? 'rgba(93,108,116,0.3)' : 'rgba(0,0,0,0.35)';
+
+  if (scrollState === ScrollState.Both) {
+    return css`
+      box-shadow: inset 6px 0 6px -6px ${shadowColor},
+        inset -6px 0 6px -6px ${shadowColor}, inset 0 6px 6px -6px ${colors[0]},
+        inset 0 -6px 6px -6px ${colors[0]};
+    `;
+  }
+
+  if (scrollState === ScrollState.Left) {
+    return css`
+      box-shadow: inset 6px 0 6px -6px ${shadowColor};
+    `;
+  }
+
+  if (scrollState === ScrollState.Right) {
+    return css`
+      box-shadow: inset -6px 0 6px -6px ${shadowColor};
+    `;
+  }
+
+  return '';
 }
 
 interface CodeProps extends SyntaxProps {
@@ -106,6 +228,13 @@ interface CodeProps extends SyntaxProps {
    * default: `true`
    * */
   multiline?: boolean;
+
+  /**
+   * When true, allows the code block to be copied to the user's clipboard
+   *
+   * default: `true`
+   */
+  copyable?: boolean;
 }
 
 type DetailedElementProps<T> = React.DetailedHTMLProps<
@@ -138,10 +267,52 @@ function Code({
   showLineNumbers = false,
   showWindowChrome = false,
   chromeTitle = '',
+  copyable = true,
   ...rest
 }: CodeProps) {
+  const scrollableMultiLine = useRef<HTMLPreElement>(null);
+  const scrollableSingleLine = useRef<HTMLDivElement>(null);
+  const [scrollState, setScrollState] = useState<ScrollState>(ScrollState.None);
+  const [copied, setCopied] = useState(false);
+  const showCopyBar = !showWindowChrome && copyable;
+
+  useEffect(() => {
+    let timeoutId: any;
+    const clipboard = new ClipboardJS('.copy-btn');
+
+    if (copied) {
+      timeoutId = setTimeout(() => {
+        setCopied(false);
+        clipboard.destroy();
+      }, 1500);
+    }
+
+    return () => clearTimeout(timeoutId);
+  }, [copied]);
+
+  useLayoutEffect(() => {
+    if (multiline) {
+      const multilineEl = scrollableMultiLine.current;
+
+      if (
+        multilineEl != null &&
+        multilineEl.scrollWidth > multilineEl.clientWidth
+      ) {
+        setScrollState(ScrollState.Right);
+      }
+    } else {
+      const singlelineEl = scrollableSingleLine.current;
+
+      if (
+        singlelineEl != null &&
+        singlelineEl.scrollWidth > singlelineEl.clientWidth
+      ) {
+        setScrollState(ScrollState.Right);
+      }
+    }
+  }, []);
+
   const wrapperStyle = css`
-    display: inline-block;
     border: 1px solid ${variantColors[variant][1]};
     border-radius: 4px;
     overflow: hidden;
@@ -155,6 +326,7 @@ function Code({
       [codeWrapperStyleWithWindowChrome]: showWindowChrome,
     },
     className,
+    getScrollShadowStyle(scrollState, variant),
   );
 
   const { content, lineCount } = useProcessedCodeSnippet(children);
@@ -169,16 +341,69 @@ function Code({
     </Syntax>
   );
 
+  function handleScroll(e: React.UIEvent) {
+    const { scrollWidth, clientWidth: elementWidth } = e.target as HTMLElement;
+    const isScrollable = scrollWidth > elementWidth;
+
+    if (isScrollable) {
+      const scrollPosition = (e.target as HTMLElement).scrollLeft;
+      const maxPosition = scrollWidth - elementWidth;
+
+      if (scrollPosition > 0 && scrollPosition < maxPosition) {
+        setScrollState(ScrollState.Both);
+      } else if (scrollPosition > 0) {
+        setScrollState(ScrollState.Left);
+      } else if (scrollPosition < maxPosition) {
+        setScrollState(ScrollState.Right);
+      }
+    }
+  }
+
+  const debounceScroll = debounce(handleScroll, 50, { leading: true });
+
   if (!multiline) {
     return (
       <div className={wrapperStyle}>
         {renderedWindowChrome}
 
         <div
-          {...(rest as DetailedElementProps<HTMLDivElement>)}
-          className={wrapperClassName}
+          className={css`
+            display: flex;
+          `}
         >
-          {renderedSyntaxComponent}
+          <div
+            {...(rest as DetailedElementProps<HTMLDivElement>)}
+            className={wrapperClassName}
+            onScroll={e => {
+              e.persist();
+              debounceScroll(e);
+            }}
+            ref={scrollableSingleLine}
+          >
+            {renderedSyntaxComponent}
+          </div>
+
+          {showCopyBar && (
+            <div
+              className={cx(
+                copyStyle,
+                singleLineCopyStyle,
+                getSidebarVariantStyle(variant),
+              )}
+            >
+              <IconButton
+                variant={variant}
+                ariaLabel="Copy"
+                className={cx(getCopyButtonStyle(variant, copied), 'copy-btn')}
+                onClick={() => {
+                  setCopied(true);
+                }}
+                data-clipboard-text={content}
+              >
+                <Icon glyph={copied ? 'Checkmark' : 'Copy'} />
+              </IconButton>
+            </div>
+          )}
         </div>
       </div>
     );
@@ -187,17 +412,43 @@ function Code({
   return (
     <div className={wrapperStyle}>
       {renderedWindowChrome}
-
-      <pre
-        {...(rest as DetailedElementProps<HTMLPreElement>)}
-        className={wrapperClassName}
+      <div
+        className={css`
+          display: flex;
+        `}
       >
-        {showLineNumbers && (
-          <LineNumbers variant={variant} lineCount={lineCount} />
-        )}
+        <pre
+          {...(rest as DetailedElementProps<HTMLPreElement>)}
+          className={wrapperClassName}
+          onScroll={e => {
+            e.persist();
+            debounceScroll(e);
+          }}
+          ref={scrollableMultiLine}
+        >
+          {showLineNumbers && (
+            <LineNumbers variant={variant} lineCount={lineCount} />
+          )}
 
-        {renderedSyntaxComponent}
-      </pre>
+          {renderedSyntaxComponent}
+        </pre>
+
+        {showCopyBar && (
+          <div className={cx(copyStyle, getSidebarVariantStyle(variant))}>
+            <IconButton
+              variant={variant}
+              ariaLabel="Copy"
+              className={cx(getCopyButtonStyle(variant, copied), 'copy-btn')}
+              onClick={() => {
+                setCopied(true);
+              }}
+              data-clipboard-text={content}
+            >
+              <Icon glyph={copied ? 'Checkmark' : 'Copy'} />
+            </IconButton>
+          </div>
+        )}
+      </div>
     </div>
   );
 }