+++ packages/mongo-nav/src/MongoNav.spec.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
-import { render, cleanup, act, fireEvent } from '@testing-library/react';
+import { render, fireEvent, cleanup, act } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import { nullableElement, Queries } from 'packages/lib/src/testHelpers';
 import { dataFixtures } from './data';
 import MongoNav from './MongoNav';
-import { NavElement } from './types';
 
 // types
 interface ExpectedElements {
@@ -31,6 +30,7 @@ describe('packages/mongo-nav', () => {
     expectedElements.currentProject = queryByTestId(
       'project-select-active-project',
     );
+    expectedElements.allProjects = queryByTestId('project-select-project-list');
     expectedElements.userMenu = queryByTestId('user-menu-trigger');
     expectedElements.userMenuLogout = queryByTestId('user-menuitem-logout');
     expectedElements.onPremUserMenu = queryByTestId('om-user-menu-trigger');
@@ -127,6 +127,19 @@ describe('packages/mongo-nav', () => {
           dataFixtures!.currentProject?.projectName as string,
         ),
       ).toBe(true);
+    });
+
+    test('projects displayed in ProjectSelect are only shown if they have the same orgId as the current project', () => {
+      const currentProject = expectedElements.currentProject;
+      fireEvent.click(currentProject as HTMLElement);
+      setExpectedElements();
+
+      const projectList = expectedElements!.allProjects;
+      const projectOptions = projectList!.querySelectorAll(
+        '[data-testid="project-option"]',
+      );
+      expect(projectOptions[0].innerHTML.includes('Demo Project 1')).toBe(true);
+      expect(projectOptions.length).toBe(1);
     });
 
     test('user is set based on data returned from fetch', () => {
@@ -331,83 +344,6 @@ describe('packages/mongo-nav', () => {
           dataFixtures!.currentOrganization!.orgName,
         ),
       ).toBe(false);
-    });
-  });
-
-  describe('when onElementClick is set', () => {
-    const onElementClick = jest.fn();
-
-    beforeEach(() =>
-      renderComponent({
-        mode: 'dev',
-        onElementClick,
-      }),
-    );
-
-    test('when Atlas is clicked', () => {
-      const atlas = (expectedElements.atlas as HTMLElement).firstChild;
-      fireEvent.click(atlas as HTMLElement);
-      expect(onElementClick).toHaveBeenCalled();
-      expect(onElementClick).toHaveBeenCalledWith(
-        NavElement.Cloud,
-        expect.anything(),
-      );
-    });
-
-    test('when Realm is clicked', () => {
-      const realm = (expectedElements.realm as HTMLElement).firstChild;
-      fireEvent.click(realm as HTMLElement);
-      expect(onElementClick).toHaveBeenCalled();
-      expect(onElementClick).toHaveBeenCalledWith(
-        NavElement.Realm,
-        expect.anything(),
-      );
-    });
-
-    test('when Charts is clicked', () => {
-      const charts = (expectedElements.charts as HTMLElement).firstChild;
-      fireEvent.click(charts as HTMLElement);
-      expect(onElementClick).toHaveBeenCalled();
-      expect(onElementClick).toHaveBeenCalledWith(
-        NavElement.Charts,
-        expect.anything(),
-      );
-    });
-
-    test('when logout is clicked', () => {
-      const userMenu = expectedElements.userMenu;
-      fireEvent.click(userMenu as HTMLElement);
-      setExpectedElements();
-      const logout = expectedElements.userMenuLogout;
-      fireEvent.click(logout as HTMLElement);
-      expect(onElementClick).toHaveBeenCalledWith(
-        NavElement.Logout,
-        expect.anything(),
-      );
-    });
-  });
-
-  describe('when onPrem and onElementClick is set', () => {
-    const onElementClick = jest.fn();
-
-    beforeEach(() =>
-      renderComponent({
-        mode: 'dev',
-        onElementClick,
-        onPrem: { enabled: true },
-      }),
-    );
-
-    test('when logout is clicked', () => {
-      const onPremUserMenu = expectedElements.onPremUserMenu;
-      fireEvent.click(onPremUserMenu as HTMLElement);
-      setExpectedElements();
-      const logout = expectedElements.onPremLogout;
-      fireEvent.click(logout as HTMLElement);
-      expect(onElementClick).toHaveBeenCalledWith(
-        NavElement.Logout,
-        expect.anything(),
-      );
     });
   });
 });